backend_llm/requirements.txt
langchain-community
langchain-core
langchain-openai
faiss-cpu
python-dotenv
openai
tiktoken
